PropertyStorage refactoring - expected gains: 0-5%
This is probably the first of more deeper dives into INTERNAL_PropertyStorage. / DependencyObject classes.

I did see memory benefits (roughly 5%), but I'm not sure if it's a conincidence or not.
It will definitely need more testing on my end to measure real (memory) benefits, but it will for sure make it easier to investigate INTERNAL_PropertyStorage. / DependencyObject  later.